Fixed Initial Grid Display
Grid would be displayed when view was loaded however the grid boundary
would not match the bounds of the game image. To fix this I had to call
the drawLines() method inside the viewDidLayoutSubviews() call instead
of viewDidLoad(). Same solution for setting the mainVC buttons back to
size when going back a page